[Refactoring] Cleanup source code unit. Move instance methods into the best matching module
[ghstack-poisoned]
