+++ src/models/scopes.rs
@@ -13,7 +13,14 @@ Copyright (c) 2022 Uber Technologies, Inc.
 
 use derive_builder::Builder;
 use getset::Getters;
+use log::trace;
 use serde_derive::Deserialize;
+
+use crate::utilities::tree_sitter_utilities::{
+  get_node_for_range, substitute_tags, PiranhaHelpers,
+};
+
+use super::{rule_store::RuleStore, source_code_unit::SourceCodeUnit};
 
 // Represents the content in the `scope_config.toml` file
 #[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default, Getters)]
@@ -37,6 +44,43 @@ pub(crate) struct ScopeQueryGenerator {
   matcher: String, // a tree-sitter query matching some enclosing AST pattern (like method or class)
   #[get = "pub"]
   generator: String, // a tree-sitter query matching the exact AST node
+}
+
+impl SourceCodeUnit {
+  /// Generate a tree-sitter based query representing the scope of the previous edit.
+  /// We generate these scope queries by matching the rules provided in `<lang>_scopes.toml`.
+  pub(crate) fn get_scope_query(
+    &self, scope_level: &str, start_byte: usize, end_byte: usize, rules_store: &mut RuleStore,
+  ) -> String {
+    let root_node = self.root_node();
+    let mut changed_node = get_node_for_range(root_node, start_byte, end_byte);
+    // Get the scope matchers for `scope_level` from the `scope_config.toml`.
+    let scope_matchers = rules_store.get_scope_query_generators(scope_level);
+
+    // Match the `scope_matcher.matcher` to the parent
+    loop {
+      trace!(
+        "Getting scope {} for node kind {}",
+        scope_level,
+        changed_node.kind()
+      );
+      for m in &scope_matchers {
+        if let Some(p_match) =
+          changed_node.get_match_for_query(self.code(), rules_store.query(m.matcher()), false)
+        {
+          // Generate the scope query for the specific context by substituting the
+          // the tags with code snippets appropriately in the `generator` query.
+          return substitute_tags(m.generator(), p_match.matches(), true);
+        }
+      }
+      if let Some(parent) = changed_node.parent() {
+        changed_node = parent;
+      } else {
+        break;
+      }
+    }
+    panic!("Could not create scope query for {scope_level:?}");
+  }
 }
 
 #[cfg(test)]
